Fix build on windows (msvc).
Also adding noexcept specifier for move constructors and assignments where possible.
This closes #2

+++ src/sdk/crypto/hash.cpp
@@ -61,7 +61,7 @@ namespace nem2_sdk {
 	typename Keccak<TMode, HashSize>::Type Keccak<TMode, HashSize>::finalize()
 	{
 		Type hash;
-		KeccakFinalize(hashInstance_, hash.data(), hash.size(), TMode{});
+		KeccakFinalize(hashInstance_, hash.data(), static_cast<uint32_t>(hash.size()), TMode{});
 		return hash;
 	}
